backend/requirements.txt
pandas
requests
feedparser
pyaudio
fastapi
uvicorn
pydantic
google-genai
yfinance
python-dotenv
numpy
scipy
